fix: トーストを transform コンテナ外に移動して表示を修正
スワイプコンテナの transform: translateX() が position: fixed の
含有ブロックになるため、HealthForm 内の toast が正しく表示されなかった。
toast state を App.tsx に移動し、transform 外のルートレベルで
レンダリングするよう修正。HealthForm には onToast コールバックを追加。
